+++ internal/infra/clients/wordofwisdom/client.go
@@ -38,9 +38,7 @@ func (c *Client) GetAuthParams(ctx context.Context) (contracts.AuthResponse, err
 
 	switch response.Type {
 	case common.MessageTypeSrvAuthResp:
-		return contracts.AuthResponse{
-			Seed: response.Payload[0], // FIXME move logic to message.go
-		}, nil
+		return contracts.AuthResponse{Seed: response.Payload[0]}, nil
 	case common.MessageTypeError:
 		return contracts.AuthResponse{}, fmt.Errorf("server returned an error: %s", strings.Join(response.Payload, ";"))
 	default:
@@ -60,9 +58,7 @@ func (c *Client) GetData(ctx context.Context, req contracts.DataRequest) (contra
 
 	switch response.Type {
 	case common.MessageTypeSrvDataResp:
-		return contracts.DataResponse{
-			MyPrecious: response.Payload[0], // FIXME move logic to message.go
-		}, nil
+		return contracts.DataResponse{MyPrecious: response.Payload[0]}, nil
 	case common.MessageTypeError:
 		return contracts.DataResponse{}, fmt.Errorf("server returned an error: %s", strings.Join(response.Payload, ";"))
 	default:
