+++ internal/infra/protocol/common/interfaces.go
@@ -2,17 +2,10 @@ package common
 
 import (
 	"context"
-	"time"
 )
-
-type RequestMeta struct {
-	RemoteAddress string
-	Time          time.Time
-}
 
 type Request struct {
 	Type    MessageType
-	Meta    RequestMeta
 	Payload []string
 }
 
@@ -22,7 +15,3 @@ type Response struct {
 }
 
 type HandlerFunc func(context.Context, Request) (Response, error)
-
-type Handler interface {
-	Handle(context.Context, Request) (Response, error)
-}
