+++ internal/infra/transport/connection/connection.go
@@ -2,19 +2,13 @@ package connection
 
 import (
 	"bufio"
-	"context"
 	"errors"
 	"fmt"
-	"io"
-	"net"
-	"slices"
 	"strings"
-	"time"
 
 	"github.com/bullgare/pow-ddos-protection/internal/infra/protocol/common"
 )
 
-// FIXME split it into 3 files.
 func New(
 	onError func(error),
 ) (*Connection, error) {
@@ -28,118 +22,10 @@ func New(
 	}, nil
 }
 
+// Connection knows how to send and receive messages over a
 type Connection struct {
 	onError   func(error)
 	delimiter byte
-}
-
-// HandleBlockingWithHandlerFunc handles all communications for one connection (meant to be used on a server side).
-// TODO we might limit connection time to not exhaust the connections.
-func (c *Connection) HandleBlockingWithHandlerFunc(
-	ctx context.Context,
-	conn net.Conn,
-	handler common.HandlerFunc,
-) {
-	defer func() { _ = conn.Close() }()
-
-	if handler == nil {
-		c.onError(errors.New("handler is required"))
-		return
-	}
-
-	r := bufio.NewReader(conn)
-	w := bufio.NewWriter(conn)
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		default:
-			raw, err := r.ReadString(c.delimiter)
-			if err == io.EOF {
-				return
-			}
-			if err != nil {
-				c.onError(fmt.Errorf("reading from conn: %w", err))
-				continue
-			}
-
-			msg, err := parseRawMessage(raw)
-			if err != nil {
-				c.onError(fmt.Errorf("parsing request: %w", err))
-				c.sendResponse(w, common.Response{Type: common.MessageTypeError, Payload: []string{err.Error()}})
-				continue
-			}
-
-			req := common.Request{
-				Type: msg.Type,
-				Meta: common.RequestMeta{
-					RemoteAddress: conn.RemoteAddr().String(),
-					Time:          time.Now(),
-				},
-				Payload: slices.Clone(msg.Payload),
-			}
-
-			resp, err := handler(ctx, req)
-			if err != nil {
-				c.onError(err)
-				c.sendResponse(w, common.Response{Type: common.MessageTypeError, Payload: []string{err.Error()}})
-				continue
-			}
-
-			c.sendResponse(w, resp)
-		}
-	}
-}
-
-// SendRequest requests a server (meant to be used on a client side).
-func (c *Connection) SendRequest(ctx context.Context, conn net.Conn, req common.Request) (common.Response, error) {
-	r := bufio.NewReader(conn)
-	w := bufio.NewWriter(conn)
-
-	msg := common.Message{
-		Version: common.MessageVersionV1,
-		Type:    req.Type,
-		Payload: slices.Clone(req.Payload),
-	}
-
-	err := c.sendMessage(w, msg)
-	if err != nil {
-		return common.Response{}, fmt.Errorf("sending request %v: %w", req, err)
-	}
-
-	select {
-	case <-ctx.Done():
-		return common.Response{}, ctx.Err()
-	default:
-		rawResp, err := r.ReadString(c.delimiter)
-		if err != nil {
-			return common.Response{}, fmt.Errorf("reading response: %w", err)
-		}
-
-		msg, err = parseRawMessage(rawResp)
-		if err != nil {
-			return common.Response{}, fmt.Errorf("parsing request: %w", err)
-		}
-
-		return common.Response{
-			Type:    msg.Type,
-			Payload: slices.Clone(msg.Payload),
-		}, nil
-	}
-}
-
-func (c *Connection) sendResponse(w *bufio.Writer, resp common.Response) {
-	msg := common.Message{
-		Version: common.MessageVersionV1,
-		Type:    resp.Type,
-		Payload: slices.Clone(resp.Payload),
-	}
-
-	err := c.sendMessage(w, msg)
-	if err != nil {
-		c.onError(err)
-	}
 }
 
 func (c *Connection) sendMessage(w *bufio.Writer, msg common.Message) error {
@@ -158,7 +44,7 @@ func (c *Connection) sendMessage(w *bufio.Writer, msg common.Message) error {
 	return nil
 }
 
-func parseRawMessage(raw string) (common.Message, error) {
+func (c *Connection) parseRawMessage(raw string) (common.Message, error) {
 	raw = strings.TrimSpace(raw)
 	chunks := strings.Split(raw, common.Separator)
 	if len(chunks) < 3 {
