--- internal/usecase/contracts/data.go
@@ -1,16 +1,0 @@
-package contracts
-
-import (
-	"time"
-)
-
-type DataRequest struct {
-	ClientRemoteAddress string
-	RequestTime         time.Time
-	Token               string
-	OriginalSeed        string
-}
-
-type DataResponse struct {
-	MyPrecious string
-}
