--- internal/usecase/contracts/interfaces.go
@@ -1,10 +1,0 @@
-package contracts
-
-import (
-	"context"
-)
-
-type ClientWordOfWisdom interface {
-	GetAuthParams(context.Context) (AuthResponse, error)
-	GetData(context.Context, DataRequest) (DataResponse, error)
-}
