+++ internal/usecase/users/user.go
@@ -1,0 +1,10 @@
+package users
+
+import (
+	"time"
+)
+
+type User struct {
+	RemoteAddress string
+	RequestTime   time.Time
+}
